server/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.4.1
    # via
    #   -r requirements/base.txt
    #   django
    #   djoser
attrs==21.2.0
    # via pytest
backcall==0.2.0
    # via
    #   -r requirements/base.txt
    #   ipython
black==21.9b0
    # via -r requirements/dev.in
certifi==2021.10.8
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.14.6
    # via
    #   -r requirements/base.txt
    #   cryptography
charset-normalizer==2.0.7
    # via
    #   -r requirements/base.txt
    #   requests
click==8.0.3
    # via black
coreapi==2.3.3
    # via
    #   -r requirements/base.txt
    #   djoser
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   drf-yasg
cryptography==35.0.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
decorator==5.1.0
    # via
    #   -r requirements/base.txt
    #   ipython
defusedxml==0.7.1
    # via
    #   -r requirements/base.txt
    #   python3-openid
    #   social-auth-core
django==3.2.8
    # via
    #   -r requirements/base.txt
    #   django-admin-display
    #   django-extensions
    #   django-filter
    #   django-jazzmin
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-admin-display==1.3.0
    # via -r requirements/base.txt
django-extensions==3.1.3
    # via -r requirements/base.txt
django-filter==21.1
    # via -r requirements/base.txt
django-jazzmin==2.4.8
    # via -r requirements/base.txt
django-json-widget==1.1.1
    # via -r requirements/base.txt
django-stubs==1.9.0
    # via -r requirements/dev.in
django-stubs-ext==0.3.1
    # via django-stubs
django-templated-mail==1.1.1
    # via
    #   -r requirements/base.txt
    #   djoser
djangorestframework==3.12.4
    # via
    #   -r requirements/base.txt
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==4.8.0
    # via
    #   -r requirements/base.txt
    #   djoser
djoser==2.1.0
    # via -r requirements/base.txt
drf-flex-fields==0.9.5
    # via -r requirements/base.txt
drf-yasg==1.20.0
    # via -r requirements/base.txt
environs==9.3.4
    # via -r requirements/base.txt
execnet==1.9.0
    # via pytest-xdist
flake8==4.0.1
    # via
    #   -r requirements/dev.in
    #   flake8-formatter-abspath
flake8-formatter-abspath==1.0.1
    # via -r requirements/dev.in
future==0.18.2
    # via
    #   -r requirements/base.txt
    #   django-json-widget
idna==3.3
    # via
    #   -r requirements/base.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements/base.txt
    #   drf-yasg
iniconfig==1.1.1
    # via pytest
ipython==7.28.0
    # via -r requirements/base.txt
isort==5.9.3
    # via -r requirements/dev.in
itypes==1.2.0
    # via
    #   -r requirements/base.txt
    #   coreapi
jedi==0.18.0
    # via
    #   -r requirements/base.txt
    #   ipython
jinja2==3.0.2
    # via
    #   -r requirements/base.txt
    #   coreschema
markupsafe==2.0.1
    # via
    #   -r requirements/base.txt
    #   jinja2
marshmallow==3.13.0
    # via
    #   -r requirements/base.txt
    #   environs
matplotlib-inline==0.1.3
    # via
    #   -r requirements/base.txt
    #   ipython
mccabe==0.6.1
    # via flake8
mypy==0.910
    # via
    #   -r requirements/dev.in
    #   django-stubs
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
oauthlib==3.1.1
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
packaging==21.0
    # via
    #   -r requirements/base.txt
    #   drf-yasg
    #   pytest
    #   pytest-sugar
parso==0.8.2
    # via
    #   -r requirements/base.txt
    #   jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via
    #   -r requirements/base.txt
    #   ipython
pickleshare==0.7.5
    # via
    #   -r requirements/base.txt
    #   ipython
platformdirs==2.4.0
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.20
    # via
    #   -r requirements/base.txt
    #   ipython
psycopg2-binary==2.9.1
    # via -r requirements/base.txt
ptyprocess==0.7.0
    # via
    #   -r requirements/base.txt
    #   pexpect
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pycodestyle==2.8.0
    # via flake8
pycparser==2.20
    # via
    #   -r requirements/base.txt
    #   cffi
pyflakes==2.4.0
    # via flake8
pygments==2.10.0
    # via
    #   -r requirements/base.txt
    #   ipython
pyjwt==2.2.0
    # via
    #   -r requirements/base.txt
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==2.4.7
    # via
    #   -r requirements/base.txt
    #   packaging
pytest==6.2.5
    # via
    #   -r requirements/dev.in
    #   pytest-django
    #   pytest-forked
    #   pytest-sugar
    #   pytest-xdist
pytest-django==4.4.0
    # via -r requirements/dev.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-sugar==0.9.4
    # via -r requirements/dev.in
pytest-xdist==2.4.0
    # via -r requirements/dev.in
python-dotenv==0.19.1
    # via
    #   -r requirements/base.txt
    #   environs
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pytz==2021.3
    # via
    #   -r requirements/base.txt
    #   django
regex==2021.10.8
    # via black
requests==2.26.0
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
ruamel.yaml==0.17.16
    # via
    #   -r requirements/base.txt
    #   drf-yasg
ruamel.yaml.clib==0.2.6
    # via
    #   -r requirements/base.txt
    #   ruamel.yaml
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via
    #   -r requirements/base.txt
    #   djoser
social-auth-core==4.1.0
    # via
    #   -r requirements/base.txt
    #   social-auth-app-django
sqlparse==0.4.2
    # via
    #   -r requirements/base.txt
    #   django
termcolor==1.1.0
    # via pytest-sugar
toml==0.10.2
    # via
    #   django-stubs
    #   mypy
    #   pytest
tomli==1.2.1
    # via black
traitlets==5.1.0
    # via
    #   -r requirements/base.txt
    #   ipython
    #   matplotlib-inline
types-pytz==2021.3.0
    # via django-stubs
types-pyyaml==5.4.11
    # via django-stubs
typing-extensions==3.10.0.2
    # via
    #   black
    #   django-stubs
    #   django-stubs-ext
    #   mypy
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.7
    # via
    #   -r requirements/base.txt
    #   requests
wcwidth==0.2.5
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
